+++ src/babel-dependencies/index.ts
@@ -1,0 +1,6 @@
+export {
+    getBabelDynamicPackageDependencies,
+    babelFindConfiguration,
+    getPresetPackageName,
+    getPluginPackageName
+} from './babel-dependencies'
