+++ src/env-utils/index.ts
@@ -16,10 +16,4 @@ export {
     TestResult
 } from './types'
 
-export {getBabelDynamicPackageDependencies, getPluginPackageName, getPresetPackageName} from './babel-dependencies'
 
-export {
-    findConfiguration,
-    findOptions,
-    FindStrategy
-} from './find-configuration'
