+++ src/find-configuration/index.ts
@@ -1,0 +1,1 @@
+export {findConfiguration, findOptions, FindStrategy, defaultGetBy} from './find-configuration'
