+++ src/jest/jest.ts
@@ -15,10 +15,9 @@ import {
 import {
     loadPackageJsonSync,
     findByName,
-    fillDependencyVersion,
-    FindStrategy,
-    findConfiguration
+    fillDependencyVersion
 } from '../env-utils'
+import {FindStrategy, findConfiguration} from '../../src/find-configuration'
 import { convertJestFormatToBitFormat } from './result-adapter'
 
 export default CreateJestTester()
