+++ src/mocha/mocha.ts
@@ -1,5 +1,6 @@
 import { TesterExtension, ExtensionApiOptions, API, ActionTesterOptions } from '../env-utils/types'
-import { loadPackageJsonSync, fillDependencyVersion, FindStrategy, findConfiguration } from '../env-utils'
+import { loadPackageJsonSync, fillDependencyVersion } from '../env-utils'
+import { FindStrategy, findConfiguration} from '../../src/find-configuration'
 import { JSONReporter } from './json-reporter'
 import Mocha, {Test} from 'mocha'
 import _get from 'lodash.get'
